+++ src/ontodia/diagram/model.ts
@@ -16,6 +16,7 @@ import {
 } from './elements';
 import { Vector } from './geometry';
 import { Graph } from './graph';
+import { CommandHistory, Command } from './history';
 
 export interface DiagramModelEvents {
     loadingStart: { source: DiagramModel };
@@ -51,7 +52,7 @@ export class DiagramModel {
 
     private classTree: FatClassModel[] = [];
 
-    constructor() {
+    constructor(readonly history: CommandHistory) {
         this.classFetching = new DataFetchingThread();
         this.linkFetching = new DataFetchingThread();
         this.propertyLabelFetching = new DataFetchingThread();
@@ -85,12 +86,6 @@ export class DiagramModel {
     isSourceAndTargetVisible(link: Link): boolean {
         return Boolean(this.sourceOf(link) && this.targetOf(link));
     }
-
-    undo() { throw new Error('History is not implemented'); }
-    redo() { throw new Error('History is not implemented'); }
-    resetHistory() { /* TODO */ }
-    initBatchCommand() { /* TODO */ }
-    storeBatchCommand() { /* TODO */ }
 
     private resetGraph() {
         if (this.graphListener) {
@@ -245,7 +240,7 @@ export class DiagramModel {
             hideUnusedLinkTypes,
         } = params;
 
-        const elementToRequestData: Element[] = [];
+        const elementIrisToRequestData: string[] = [];
         const usedLinkTypes: { [typeId: string]: FatLinkType } = {};
 
         const normalizedCells = layoutData.cells.map(normalizeImportedCell);
@@ -257,7 +252,7 @@ export class DiagramModel {
                 const element = new Element({id, data, position, size, expanded: isExpanded, group});
                 this.graph.addElement(element);
                 if (!template) {
-                    elementToRequestData.push(element);
+                    elementIrisToRequestData.push(element.iri);
                 }
             }
         }
@@ -267,7 +262,7 @@ export class DiagramModel {
                 const {id, typeId, source, target, vertices} = cell;
                 const linkType = this.createLinkType(typeId);
                 usedLinkTypes[linkType.id] = linkType;
-                const link = this.graph.createLink({
+                const link = this.createLink({
                     data: {
                         linkTypeId: typeId,
                         sourceId: source.id,
@@ -283,7 +278,7 @@ export class DiagramModel {
         }
 
         this.subscribeGraph();
-        this.requestElementData(elementToRequestData);
+        this.requestElementData(elementIrisToRequestData);
 
         if (hideUnusedLinkTypes && params.allLinkTypes) {
             this.hideUnusedLinkTypes(params.allLinkTypes, usedLinkTypes);
@@ -306,11 +301,11 @@ export class DiagramModel {
         }
     }
 
-    requestElementData(elements: Element[]): Promise<void> {
-        if (elements.length === 0) {
+    requestElementData(elementIris: ReadonlyArray<string>): Promise<void> {
+        if (elementIris.length === 0) {
             return Promise.resolve();
         }
-        return this.dataProvider.elementInfo({elementIds: elements.map(e => e.iri)})
+        return this.dataProvider.elementInfo({elementIds: [...elementIris]})
             .then(models => this.onElementInfoLoaded(models))
             .catch(err => {
                 console.error(err);
@@ -404,12 +399,20 @@ export class DiagramModel {
             ? placeholderTemplateFromIri(elementIri) : elementIriOrModel;
         data = {...data, id: data.id};
         const element = new Element({id: `element_${generate64BitID()}`, data, group});
-        this.graph.addElement(element);
+        this.history.execute(
+            addElement(this.graph, element, [])
+        );
+
         return element;
     }
 
     removeElement(elementId: string) {
-        this.graph.removeElement(elementId);
+        const element = this.getElement(elementId);
+        if (element) {
+            this.history.execute(
+                removeElement(this.graph, element)
+            );
+        }
     }
 
     createLinkType(linkTypeId: string): FatLinkType {
@@ -464,12 +467,10 @@ export class DiagramModel {
     }
 
     private onLinkInfoLoaded(links: LinkModel[]) {
-        this.initBatchCommand();
         for (const linkModel of links) {
             const linkType = this.createLinkType(linkModel.linkTypeId);
             this.createLinks(linkModel, linkType);
         }
-        this.storeBatchCommand();
     }
 
     private createLinks(linkModel: LinkModel, linkType: FatLinkType) {
@@ -480,9 +481,38 @@ export class DiagramModel {
         for (const source of sources) {
             for (const target of targets) {
                 const data = {...linkModel, sourceId: source.id, targetId: target.id};
-                this.graph.createLink({data, linkType});
+                this.createLink({data, linkType});
             }
         }
+    }
+
+    private createLink(params: {
+        data: LinkModel;
+        linkType: FatLinkType;
+        vertices?: ReadonlyArray<Vector>;
+    }): Link {
+        const {data, linkType, vertices} = params;
+        if (data.linkTypeId !== linkType.id) {
+            throw new Error('linkTypeId must match linkType.id');
+        }
+
+        const existingLink = this.findLink(data);
+        if (existingLink) {
+            existingLink.setLayoutOnly(false);
+            return existingLink;
+        }
+
+        const shouldBeVisible = linkType.visible
+            && this.getElement(data.sourceId)
+            && this.getElement(data.targetId);
+
+        if (!shouldBeVisible) {
+            return undefined;
+        }
+
+        const link = new Link({id: `link_${generate64BitID()}`, data, vertices});
+        this.graph.addLink(link);
+        return link;
     }
 }
 
@@ -492,11 +522,37 @@ export interface LinkTypeOptions {
     showLabel?: boolean;
 }
 
+function addElement(graph: Graph, element: Element, connectedLinks: ReadonlyArray<Link>): Command {
+    return Command.create('Add element', () => {
+        graph.addElement(element);
+        for (const link of connectedLinks) {
+            if (graph.getLink(link.id)) { continue; }
+            graph.addLink(link);
+        }
+        return removeElement(graph, element);
+    });
+}
+
+function removeElement(graph: Graph, element: Element): Command {
+    return Command.create('Remove element', () => {
+        const connectedLinks = [...element.links];
+        graph.removeElement(element.id);
+        return addElement(graph, element, connectedLinks);
+    });
+}
+
+export function restoreLinksBetweenElements(model: DiagramModel, elementIris: ReadonlyArray<string>): Command {
+    return Command.effect('Restore links between elements', () => {
+        model.requestElementData(elementIris);
+        model.requestLinksOfType();
+    });
+}
+
 function placeholderTemplateFromIri(iri: string): ElementModel {
     return {
         id: iri,
         types: [],
-        label: {values: [{lang: '', text: uri2name(iri)}]},
+        label: {values: []},
         properties: {},
     };
 }